requirements.txt
fastapi
uvicorn
pyodbc
pandas
matplotlib
azure-identity
openai
sqlalchemy
python-dotenv
jinja2
seaborn
plotly